Decompose the queries in list research contexts
